late war karl moment
The Austro-Hungarian Empire was wrecked by inner turmoil in the final years of the war, with escalating tension between ethnic groups. As part of his Fourteen Points, U.S. President Woodrow Wilson demanded that the Empire allow for autonomy and self-determination of its peoples. In response, Charles agreed to reconvene the Imperial Parliament and allow for the creation of a confederation with each national group exercising self-governance. However, the ethnic groups fought for full autonomy as separate nations, as they were now determined to become independent from Vienna at the earliest possible moment.